+++ src/applications/personalization/profile360/vet360/containers/Vet360TransactionReporter.jsx
@@ -7,7 +7,7 @@ import scrollToTop from '../../../../../platform/utilities/ui/scrollToTop';
 import {
   selectVet360FailedTransactions,
   selectMostRecentErroredTransaction
-} from '../../selectors';
+} from '../selectors';
 
 import {
   clearTransaction
